\name{expressionSpeCondExample}
\Rdversion{1.1}
\alias{expressionSpeCondExample}
\docType{data}
\title{
The expression matrix example used in the SpeCond package
}
\description{
\code{expressionSpeCondExample} is expression value matrix (log2) used as an example for the SpeCond package.  This is a subset of a new analysis of the Su et al, 2008 data. The columns are human tissues, the rows are probeset IDs.
}
\usage{data(expressionSpeCondExample)}
\format{ A matrix of 220 rows and 32 columns}
\source{
Su et al, PNAS, 2004, 'A gene atlas of the mouse and human protein-encoding transcriptomes'
}
\author{Florence Cavalli, florence@ebi.ac.uk}
\examples{
data(expressionSpeCondExample)
}
\keyword{datasets}
